Add files via upload
S t a t e m e n t :

The FFMU is a research organization consisting of about 70 company members
in the region of Baden-Württemberg, who are working in the field of fine 
and precise machining, sensors/transducers and microelectronics. The FFMU 
is mainly supported by the government and the ministery of science and 
technology (BMFT). The results of our research are mainly for public domain. 
The function of the FFMU is to transfer 'know-how' and technology in their 
research fields to the industry and to their member companies.